routes moved out of components floder
