--- .vite/deps_temp_6515f218/package.json
@@ -1,3 +1,0 @@
-{
-  "type": "module"
-}
