+++ src/App.jsx
@@ -1,15 +1,16 @@
-import ContactPage from './components/Routes/ContactPage';
-import HomePage from './components/Routes/HomePage';
-import ProductsList from './components/Routes/ProductsList';
-import About from './components/Routes/About';
-import Services from './components/Routes/services';
-import TermsOfUse from './components/Routes/TermsOfUse';
-import ProductPage from './components/Routes/Products/ProductsPage';
-import ProductDetailPage from './components/Routes/Products/ProductDetailPage';
-import PrecisionFarming from './components/Routes/PrecisionFarming';
-import ScrollToTop from './components/animations/ScrollToTop';
-import FAQ from './components/Routes/FAQ';
-import Cookies from './components/Routes/Cookies';
+import ContactPage from '/src/routes/ContactPage.jsx';
+import HomePage from '/src/routes/HomePage';
+import ProductsList from '/src/routes/ProductsList';
+import About from '/src/routes/About.jsx';
+import Services from '/src/routes/services/Services.jsx';
+import TermsOfUse from '/src/routes/TermsOfUse';
+import ProductPage from '/src/routes/Products/ProductsPage';
+import ProductDetailPage from '/src/routes/Products/ProductDetailPage';
+import PrecisionFarming from '/src/routes/PrecisionFarming';
+import ScrollToTop from '/src/components/animations/ScrollToTop';
+import FAQ from '/src/routes/FAQ';
+import Cookies from '/src/routes/Cookies';
+
 
 import { BrowserRouter as Router, Routes, Route } from "react-router-dom";
