+++ src/routes/ContactPage.jsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import NavBar from '../custom/NavBar'
+import NavBar from 'src/components/custom/NavBar'
 import ContactUs from './Contact/ContactUs'
 import ContactFooter from './Contact/ContactFooter'
 import Footer from '../custom/Footer'
