+++ src/routes/HomePage.jsx
@@ -1,7 +1,7 @@
-import NavBar from "../custom/NavBar";
+import NavBar from "src/components/custom/NavBar";
 import Presentation from "./Home/Content/Presentation";
-import PriceTable from "../custom/PriceTable";
-import Footer from "../custom/Footer";
+import PriceTable from "src/components/custom/PriceTable";
+import Footer from "src/components/custom/Footer";
 import Cont0 from "./Home/Content/Cont0";
 import Cont2 from "./Home/Content/Cont2";
 import Oferta from "./Other/Notused/Offer";
