+++ src/routes/PrecisionFarming.jsx
