+++ src/routes/PrecisionFarming/Cont3.jsx
