+++ src/routes/PrecisionFarming/Cont6.jsx
