+++ src/routes/PrecisionFarming/Map2.jsx
