+++ src/routes/PrecisionFarming/Map3.jsx
