+++ src/routes/PrecisionFarming/Map_time_slider.jsx
