+++ src/routes/PrecisionFarming/SimpleMapComponent.jsx
