+++ src/routes/ProductsList.jsx
@@ -1,7 +1,7 @@
-import NavBar from "../custom/NavBar";
-import Footer from "../custom/Footer";
+import NavBar from "src/components/custom/NavBar";
+import Footer from "src/components/custom/Footer";
 import ContactUs from "./Contact/ContactUs";
-import PriceTable from "../custom/PriceTable";
+import PriceTable from "src/components/custom/PriceTable";
