--- test_file.txt
@@ -1,1 +1,0 @@
-Test file XDDD
